More wip on refactoring preferences
