+++ lib/src/model/analysis/analysis_preferences.dart
@@ -1,31 +1,24 @@
-import 'dart:convert';
-
-import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:lichess_mobile/src/db/shared_preferences.dart';
 import 'package:lichess_mobile/src/model/engine/evaluation_service.dart';
+import 'package:lichess_mobile/src/model/settings/preferences.dart' as pref;
+import 'package:lichess_mobile/src/model/settings/preferences_storage.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-part 'analysis_preferences.freezed.dart';
 part 'analysis_preferences.g.dart';
 
-@Riverpod(keepAlive: true)
-class AnalysisPreferences extends _$AnalysisPreferences {
-  static const prefKey = 'preferences.analysis';
+@riverpod
+class AnalysisPreferences extends _$AnalysisPreferences
+    with PreferencesStorage<pref.Analysis> {
+  // ignore: avoid_public_notifier_properties
+  @override
+  pref.Category<pref.Analysis> get categoryKey => pref.Category.analysis;
 
   @override
-  AnalysisPrefState build() {
-    final prefs = ref.watch(sharedPreferencesProvider);
-
-    final stored = prefs.getString(prefKey);
-    return stored != null
-        ? AnalysisPrefState.fromJson(
-            jsonDecode(stored) as Map<String, dynamic>,
-          )
-        : AnalysisPrefState.defaults;
+  pref.Analysis build() {
+    return fetch();
   }
 
   Future<void> toggleEnableLocalEvaluation() {
-    return _save(
+    return save(
       state.copyWith(
         enableLocalEvaluation: !state.enableLocalEvaluation,
       ),
@@ -33,7 +26,7 @@ class AnalysisPreferences extends _$AnalysisPreferences {
   }
 
   Future<void> toggleShowEvaluationGauge() {
-    return _save(
+    return save(
       state.copyWith(
         showEvaluationGauge: !state.showEvaluationGauge,
       ),
@@ -41,7 +34,7 @@ class AnalysisPreferences extends _$AnalysisPreferences {
   }
 
   Future<void> toggleAnnotations() {
-    return _save(
+    return save(
       state.copyWith(
         showAnnotations: !state.showAnnotations,
       ),
@@ -49,7 +42,7 @@ class AnalysisPreferences extends _$AnalysisPreferences {
   }
 
   Future<void> togglePgnComments() {
-    return _save(
+    return save(
       state.copyWith(
         showPgnComments: !state.showPgnComments,
       ),
@@ -57,7 +50,7 @@ class AnalysisPreferences extends _$AnalysisPreferences {
   }
 
   Future<void> toggleShowBestMoveArrow() {
-    return _save(
+    return save(
       state.copyWith(
         showBestMoveArrow: !state.showBestMoveArrow,
       ),
@@ -66,7 +59,7 @@ class AnalysisPreferences extends _$AnalysisPreferences {
 
   Future<void> setNumEvalLines(int numEvalLines) {
     assert(numEvalLines >= 1 && numEvalLines <= 3);
-    return _save(
+    return save(
       state.copyWith(
         numEvalLines: numEvalLines,
       ),
@@ -75,53 +68,10 @@ class AnalysisPreferences extends _$AnalysisPreferences {
 
   Future<void> setEngineCores(int numEngineCores) {
     assert(numEngineCores >= 1 && numEngineCores <= maxEngineCores);
-    return _save(
+    return save(
       state.copyWith(
         numEngineCores: numEngineCores,
       ),
     );
-  }
-
-  Future<void> _save(AnalysisPrefState newState) async {
-    final prefs = ref.read(sharedPreferencesProvider);
-    await prefs.setString(
-      prefKey,
-      jsonEncode(newState.toJson()),
-    );
-    state = newState;
-  }
-}
-
-@Freezed(fromJson: true, toJson: true)
-class AnalysisPrefState with _$AnalysisPrefState {
-  const AnalysisPrefState._();
-
-  const factory AnalysisPrefState({
-    required bool enableLocalEvaluation,
-    required bool showEvaluationGauge,
-    required bool showBestMoveArrow,
-    required bool showAnnotations,
-    required bool showPgnComments,
-    @Assert('numEvalLines >= 1 && numEvalLines <= 3') required int numEvalLines,
-    @Assert('numEngineCores >= 1 && numEngineCores <= maxEngineCores')
-    required int numEngineCores,
-  }) = _AnalysisPrefState;
-
-  static final defaults = AnalysisPrefState(
-    enableLocalEvaluation: true,
-    showEvaluationGauge: true,
-    showBestMoveArrow: true,
-    showAnnotations: true,
-    showPgnComments: true,
-    numEvalLines: 2,
-    numEngineCores: defaultEngineCores,
-  );
-
-  factory AnalysisPrefState.fromJson(Map<String, dynamic> json) {
-    try {
-      return _$AnalysisPrefStateFromJson(json);
-    } catch (_) {
-      return defaults;
-    }
   }
 }
