--- lib/src/model/settings/sound_theme.dart
@@ -1,12 +1,0 @@
-enum SoundTheme {
-  standard('Standard'),
-  piano('Piano'),
-  nes('NES'),
-  sfx('SFX'),
-  futuristic('Futuristic'),
-  lisp('Lisp');
-
-  final String label;
-
-  const SoundTheme(this.label);
-}
