+++ test/model/common/service/fake_sound_service.dart
@@ -1,5 +1,5 @@
 import 'package:lichess_mobile/src/model/common/service/sound_service.dart';
-import 'package:lichess_mobile/src/model/settings/sound_theme.dart';
+import 'package:lichess_mobile/src/model/settings/preferences.dart';
 
 class FakeSoundService implements SoundService {
   @override
