+++ test/test_provider_scope.dart
@@ -20,7 +20,7 @@ import 'package:lichess_mobile/src/model/common/http.dart';
 import 'package:lichess_mobile/src/model/common/service/sound_service.dart';
 import 'package:lichess_mobile/src/model/common/socket.dart';
 import 'package:lichess_mobile/src/model/notifications/notification_service.dart';
-import 'package:lichess_mobile/src/model/settings/board_preferences.dart';
+import 'package:lichess_mobile/src/model/settings/preferences.dart';
 import 'package:lichess_mobile/src/utils/connectivity.dart';
 import 'package:logging/logging.dart';
 import 'package:package_info_plus/package_info_plus.dart';
@@ -103,7 +103,7 @@ Future<Widget> makeProviderScope(
         {
           // disable piece animation to simplify tests
           'preferences.board': jsonEncode(
-            BoardPrefs.defaults
+            Board.defaults
                 .copyWith(
                   pieceAnimation: false,
                 )
